Moved PhotoSettings back to Swing
